requirements.txt
arxiv==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf2==3.0.1
beautifulsoup4==4.12.2
markdown==3.5.1
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.0
